# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API для проверки клиентских CSV/XLS/XLSX, подтверждённого импорта контактов и подготовки рассылок.\nЧерновик хранится в браузере; PostgreSQL изменяет финальный атомарный импорт и явное ручное сохранение контакта.",
        "title": "Task1 Client Data API",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
        "/api/contacts": {
            "get": {
                "description": "Возвращает 25 подтверждённых контактов PostgreSQL и общее количество совпадений. Поиск выполняется по телефону, ФИО, email, скидке и UID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Contacts"
                ],
                "summary": "Получить страницу контактов",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Номер страницы, начиная с 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Подстрока для поиска по всем видимым полям",
                        "name": "q",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ContactPage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/contacts/{uid}": {
            "put": {
                "description": "Нормализует поля и обновляет контакт, только если его updatedAt не изменился. Телефон остаётся уникальным.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Contacts"
                ],
                "summary": "Изменить контакт",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID контакта",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Изменяемые поля и текущая версия",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ContactUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Contact"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/imports/commit": {
            "post": {
                "description": "Повторно проверяет черновик и одной транзакцией сохраняет файл, строки, контакты и решения конфликтов.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Imports"
                ],
                "summary": "Импортировать подтверждённые контакты",
                "parameters": [
                    {
                        "description": "Черновик и решения всех конфликтов",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CommitImportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ImportCommitResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/imports/preview": {
            "post": {
                "description": "Возвращает новые, совпадающие и конфликтующие контакты. Операция выполняет только чтение PostgreSQL.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Imports"
                ],
                "summary": "Предпросмотр последствий импорта",
                "parameters": [
                    {
                        "description": "Проверенный локальный черновик",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.PreviewImportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ImportPreviewResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/imports/validate": {
            "post": {
                "description": "Нормализует строки после редактирования и возвращает новую диагностику. PostgreSQL не изменяется.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Imports"
                ],
                "summary": "Повторно проверить локальный черновик",
                "parameters": [
                    {
                        "description": "Локальный черновик",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ValidateImportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ImportValidationResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.CommitImportRequest": {
            "type": "object",
            "properties": {
                "decisions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ImportDecision"
                    }
                },
                "draft": {
                    "$ref": "#/definitions/models.ImportDraft"
                }
            }
        },
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Некорректный запрос"
                }
            }
        },
...
        "handlers.PreviewImportRequest": {
            "type": "object",
            "properties": {
                "draft": {
                    "$ref": "#/definitions/models.ImportDraft"
                }
            }
        },
...
        "handlers.ValidateImportRequest": {
            "type": "object",
            "properties": {
                "draft": {
                    "$ref": "#/definitions/models.ImportDraft"
                }
            }
        },
        "models.ConflictAction": {
            "type": "string",
            "enum": [
                "skip",
                "replace",
                "merge"
            ],
            "x-enum-varnames": [
                "ConflictActionSkip",
                "ConflictActionReplace",
                "ConflictActionMerge"
            ]
        },
        "models.ConflictInfo": {
            "type": "object",
            "properties": {
                "actions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ConflictAction"
                    }
                },
                "differences": {
                    "description": "Отличные поля",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "existing": {
                    "description": "То, что уже лежит в базе",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "incoming": {
                    "description": "То, что прислал фронт",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "phone": {
                    "type": "string",
                    "example": "+79991234567"
                },
                "row": {
                    "type": "integer",
                    "example": 4
                },
                "version": {
                    "type": "string",
                    "example": "2026-07-16T08:30:00.123456Z"
                }
            }
        },
        "models.Contact": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "discount": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "uid": {
                    "description": "UID - публичный UUID контакта, который генерирует PostgreSQL.",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.ContactPage": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Contact"
                    }
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "pageSize": {
                    "type": "integer",
                    "example": 25
                },
                "query": {
                    "type": "string",
                    "example": "+7999"
                },
                "total": {
                    "type": "integer",
                    "example": 106
                },
                "totalPages": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
        "models.ContactUpdateRequest": {
            "type": "object",
            "properties": {
                "discount": {
                    "type": "string",
                    "example": "10"
                },
                "email": {
                    "type": "string",
                    "example": "client@example.com"
                },
                "name": {
                    "type": "string",
                    "example": "Иванов Иван"
                },
                "phone": {
                    "type": "string",
                    "example": "+79991234567"
                },
                "version": {
                    "type": "string",
                    "example": "2026-07-16T08:30:00.123456Z"
                }
            }
        },
        "models.ImportCommitResult": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "integer",
                    "example": 90
                },
                "importId": {
                    "type": "string"
                },
                "matched": {
                    "type": "integer",
                    "example": 10
                },
                "merged": {
                    "type": "integer",
                    "example": 1
                },
                "replaced": {
                    "type": "integer",
                    "example": 2
                },
                "skipped": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "models.ImportDecision": {
            "type": "object",
            "properties": {
                "action": {
                    "enum": [
                        "skip",
                        "replace",
                        "merge"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.ConflictAction"
                        }
                    ],
                    "example": "merge"
                },
                "phone": {
                    "type": "string",
                    "example": "+79991234567"
                },
                "version": {
                    "type": "string",
                    "example": "2026-07-16T08:30:00.123456Z"
                }
            }
        },
        "models.ImportDraft": {
            "type": "object",
            "properties": {
                "detectedMimeType": {
                    "type": "string"
                },
                "emptyRowCount": {
                    "type": "integer"
                },
                "encoding": {
                    "type": "string",
                    "example": "UTF-8"
                },
                "format": {
                    "type": "string",
                    "enum": [
                        "csv",
                        "xls",
                        "xlsx"
                    ],
                    "example": "xlsx"
                },
                "headerRow": {
                    "type": "integer",
                    "example": 1
                },
                "headers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "importId": {
                    "type": "string",
                    "example": "2f656bc0-6227-49d3-9d09-b2d59bd21c52"
                },
                "mimeType": {
                    "type": "string"
                },
                "originalFilename": {
                    "type": "string",
                    "example": "clients.xlsx"
                },
                "rows": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ImportRow"
                    }
                },
                "sheetName": {
                    "type": "string",
                    "example": "Клиенты"
                },
                "sheets": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "size": {
                    "type": "integer",
                    "example": 18432
                },
                "skippedRowCount": {
                    "type": "integer"
                }
            }
        },
        "models.ImportPreviewResult": {
            "type": "object",
            "properties": {
                "conflictCount": {
                    "type": "integer",
                    "example": 4
                },
                "conflicts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ConflictInfo"
                    }
                },
                "matchedCount": {
                    "type": "integer",
                    "example": 10
                },
                "newCount": {
                    "type": "integer",
                    "example": 90
                },
                "skippedCount": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
...
        "models.ImportValidationResult": {
            "type": "object",
            "properties": {
                "detectedPhoneColumn": {
                    "type": "string",
                    "example": "Телефон"
                },
                "draft": {
                    "$ref": "#/definitions/models.ImportDraft"
                },
                "invalidRows": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.InvalidRow"
                    }
                },
                "previewRows": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "string"
                        }
                    }
                },
                "stats": {
                    "$ref": "#/definitions/models.ProcessingStats"
                },
                "warnings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ProcessingWarning"
                    }
                }
            }
        },
        "models.InvalidRow": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ProcessingWarning"
                    }
                },
                "row": {
                    "type": "integer",
                    "example": 4
                },
                "values": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
        "models.ProcessingStats": {
            "type": "object",
            "properties": {
                "columnCount": {
                    "type": "integer",
                    "example": 4
                },
                "emptyRowCount": {
                    "type": "integer",
                    "example": 1
                },
                "invalidRowCount": {
                    "type": "integer",
                    "example": 3
                },
                "rowCount": {
                    "type": "integer",
                    "example": 100
                },
                "skippedRowCount": {
                    "type": "integer",
                    "example": 1
                },
                "validRowCount": {
                    "type": "integer",
                    "example": 96
                },
                "warningCount": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
...
    },
...
}